+++ opencompass/configs/models/hf_tigerbot_exllama.py
@@ -1,4 +1,5 @@
-from opencompass.models import  TigerbotExllama
+
+from opencompass.models import ExllamaCausalLM
 
 _meta_template = dict(
     round=[
@@ -9,7 +10,7 @@ _meta_template = dict(
 
 models = [
     dict(
-        type= TigerbotExllama,
+        type= ExllamaCausalLM,
         abbr='tigerbot',
         path="/mnt/nfs/algo/intern/yuwang/Tigerbot_AutoGPTQ/tigerbot_13b/tigerbot_13b_chat_4bit_c4_128g_no_act",
         tokenizer_path='/mnt/nfs/algo/intern/yuwang/Tigerbot_AutoGPTQ/tigerbot_13b/tigerbot_13b_chat_4bit_c4_128g_no_act',
@@ -25,7 +26,6 @@ models = [
         max_out_len=100,
         max_seq_len=2048,
         batch_size=4,
-        meta_template=_meta_template,
         batch_padding=True,
         run_cfg=dict(num_gpus=1, num_procs=1),
     )
