+++ opencompass/opencompass/models/huggingface.py
@@ -8,10 +8,173 @@ from opencompass.models.base import BaseModel
 from opencompass.registry import MODELS
 from opencompass.utils.logging import get_logger
 from opencompass.utils.prompt import PromptList
+
+from pathlib import Path
+from typing import Any, Dict, Optional, Union
+
+from exllama_lib.model import ExLlama, ExLlamaCache, ExLlamaConfig
+from torch.nn import CrossEntropyLoss
+from transformers import (
+    GenerationConfig,
+    PretrainedConfig,
+    PreTrainedModel,
+)
 from transformers import AutoTokenizer, LlamaTokenizer, AutoModelForCausalLM, LlamaForCausalLM, AutoConfig
+from transformers.modeling_outputs import CausalLMOutputWithPast
 
 PromptType = Union[PromptList, str]
 
+os.environ["TOKENIZERS_PARALLELISM"] = "false"
+
+tok_ins = "\n\n### Instruction:\n"
+tok_res = "\n\n### Response:\n"
+prompt_input = tok_ins + "{instruction}" + tok_res
+
+
+class ExllamaHF(PreTrainedModel):
+    def __init__(self, config: ExLlamaConfig):
+        super().__init__(PretrainedConfig())
+        self.ex_config = config
+        self.ex_model = ExLlama(self.ex_config)
+        self.generation_config = GenerationConfig()
+        self.lora = None
+
+        self.ex_cache = ExLlamaCache(self.ex_model)
+        self.past_seq = None
+
+    def _validate_model_class(self):
+        pass
+
+    def _validate_model_kwargs(self, model_kwargs: Dict[str, Any]):
+        pass
+
+    def prepare_inputs_for_generation(self, input_ids, **kwargs):
+        return {"input_ids": input_ids, **kwargs}
+
+    @property
+    def device(self) -> torch.device:
+        return torch.device(0)
+
+    def __call__(self, *args, **kwargs):
+        use_cache = kwargs.get("use_cache", True)
+        labels = kwargs.get("labels", None)
+        past_key_values = kwargs.get("past_key_values", None)
+
+        if len(args) > 0:
+            input_ids = args[0]
+            is_negative = True
+            past_seq = self.past_seq_negative
+            ex_cache = self.ex_cache_negative
+        else:
+            input_ids = kwargs["input_ids"]
+            is_negative = False
+            past_seq = self.past_seq
+            ex_cache = self.ex_cache
+
+        seq = input_ids[0].tolist()
+        if is_negative and past_key_values is not None:
+            seq = past_key_values + seq
+
+        seq_tensor = torch.tensor(seq)
+
+        # Make the forward call
+        if labels is None:
+            if past_seq is None or not torch.equal(
+                    past_seq, seq_tensor[:-1]
+            ):
+                ex_cache.current_seq_len = 0
+                self.ex_model.forward(
+                    torch.tensor([seq[:-1]], dtype=torch.long),
+                    ex_cache,
+                    preprocess_only=True,
+                    lora=self.lora,
+                )
+
+            logits = self.ex_model.forward(
+                torch.tensor([seq[-1:]], dtype=torch.long),
+                ex_cache,
+                lora=self.lora,
+            ).to(input_ids.device)
+        else:
+            ex_cache.current_seq_len = 0
+            logits = self.ex_model.forward(
+                torch.tensor([seq], dtype=torch.long),
+                ex_cache,
+                last_id_only=False,
+                lora=self.lora,
+            )
+
+        if is_negative:
+            self.past_seq_negative = seq_tensor
+        else:
+            self.past_seq = seq_tensor
+
+        loss = None
+        if labels is not None:
+            # Shift so that tokens < n predict n
+            shift_logits = logits[..., :-1, :].contiguous()
+            shift_labels = labels[..., 1:].contiguous()
+            # Flatten the tokens
+            loss_fct = CrossEntropyLoss()
+            shift_logits = shift_logits.view(-1, logits.shape[-1])
+            shift_labels = shift_labels.view(-1)
+            # Enable model parallelism
+            shift_labels = shift_labels.to(shift_logits.device)
+            loss = loss_fct(shift_logits, shift_labels)
+
+        return CausalLMOutputWithPast(
+            logits=logits,
+            past_key_values=seq if use_cache else None,
+            loss=loss,
+        )
+
+    @classmethod
+    def from_pretrained(
+            cls,
+            pretrained_model_name_or_path: Optional[
+                Union[str, os.PathLike]
+            ],
+            *model_args,
+            **kwargs,
+    ):
+        assert (
+                len(model_args) == 0 and len(kwargs) == 0
+        ), "extra args is currently not supported"
+        if isinstance(pretrained_model_name_or_path, str):
+            pretrained_model_name_or_path = Path(
+                pretrained_model_name_or_path
+            )
+
+        config = ExLlamaConfig(
+            pretrained_model_name_or_path / "config.json"
+        )
+
+        weight_path = None
+        for ext in [".safetensors", ".pt", ".bin"]:
+            found = list(pretrained_model_name_or_path.glob(f"*{ext}"))
+            if len(found) > 0:
+                weight_path = found[-1]
+                break
+        assert (
+                weight_path is not None
+        ), f'could not find weight in "{pretrained_model_name_or_path}"'
+
+        config.model_path = str(weight_path)
+        config.max_seq_len = 2048
+        config.compress_pos_emb = 1
+
+        if torch.version.hip:
+            config.rmsnorm_no_half2 = True
+            config.rope_no_half2 = True
+            config.matmul_no_half2 = True
+            config.silu_no_half2 = True
+
+        # This slowes down a bit but align better with autogptq generation.
+        # TODO: Should give user choice to tune the exllama config
+        # config.fused_attn = False
+        # config.fused_mlp_thd = 0
+
+        return ExllamaHF(config)
 
 @MODELS.register_module()
 class HuggingFace(BaseModel):
@@ -183,8 +346,9 @@ class HuggingFace(BaseModel):
             for k in tokens if k in ['input_ids', 'attention_mask']
         }
         # step-2: conduct model forward to generate output
-        outputs = self.model.generate(**tokens, max_new_tokens=max_out_len)
-
+        outputs = self.model.generate(**tokens, max_new_tokens=max_out_len,
+                                      eos_token_id=self.tokenizer.eos_token_id, pad_token_id=self.tokenizer.pad_token_id)
+        # self.model.generate(**tokens, max_new_tokens=max_out_len)
         if not self.extract_pred_after_decode:
             outputs = outputs[:, tokens['input_ids'].shape[1]:]
 
@@ -216,7 +380,9 @@ class HuggingFace(BaseModel):
                                               max_out_len)['input_ids']
         input_ids = torch.tensor(input_ids, device=self.model.device)
         outputs = self.model.generate(input_ids=input_ids,
-                                      max_new_tokens=max_out_len)
+                                      max_new_tokens=max_out_len,
+                                      eos_token_id=self.tokenizer.eos_token_id, 
+                                      pad_token_id=self.tokenizer.pad_token_id)
 
         if not self.extract_pred_after_decode:
             outputs = outputs[:, input_ids.shape[1]:]
@@ -366,6 +532,7 @@ class HuggingFaceCausalLM(HuggingFace):
                     path: str,
                     model_kwargs: dict,
                     peft_path: Optional[str] = None):
+
         model_kwargs.setdefault('torch_dtype', torch.float16)
         if "llama" in self.config.model_type.lower():
             self.model = LlamaForCausalLM.from_pretrained(path, **model_kwargs)
@@ -380,3 +547,31 @@ class HuggingFaceCausalLM(HuggingFace):
                                                    peft_path,
                                                    is_trainable=False)
         self.model.eval()
+
+
+@MODELS.register_module()
+class GPTQCausalLM(HuggingFace):
+
+    def _load_model(self,
+                    path: str,
+                    model_kwargs: dict,
+                    peft_path: str = None):
+        from auto_gptq import AutoGPTQForCausalLM, BaseQuantizeConfig
+ 
+        quantize_config = BaseQuantizeConfig.from_pretrained(path)
+        self.model = AutoGPTQForCausalLM.from_quantized(path, quantize_config=quantize_config, **model_kwargs)
+
+        self.model.eval()
+
+
+@MODELS.register_module()
+class ExllamaCausalLM(HuggingFace):
+
+    def _load_model(self,
+                    path: str,
+                    model_kwargs: dict,
+                    peft_path: str = None):
+        
+        self.model = ExllamaHF.from_pretrained(path)
+        self.model.eval()
+
